refactor: move grid related logic to own manager
